rewrote save() and save_image() functions in io to handle tifs with metadata. This is not covered in current unittests but I haven't written more yet.
