+++ holopy/core/__init__.py
@@ -42,6 +42,6 @@ from .marray import (Marray, Image, Schema, ImageSchema, Volume,
                     VolumeSchema, VectorGrid, VectorGridSchema,
                     subimage)
 from .metadata import Optics, Grid, Angles, UnevenGrid, interpret_args
-from .io import load, load_image, save
+from .io import load, load_image, save, save_image
 from . import process
 from . import helpers
