+++ holopy/core/errors.py
@@ -34,3 +34,7 @@ class BadImage(Exception):
 
 class UnspecifiedPosition(Exception):
         pass
+
+class NoMetadata(Exception):
+    def __str__(self):
+        return "File without metadata detected. To load raw images, use hp.load_image()"
